refactor: split skills cli runner
