+++ extensions/skills/src/utils/exec-async.ts
@@ -1,4 +1,4 @@
-import { exec } from "node:child_process";
+import { execFile } from "node:child_process";
 import { promisify } from "node:util";
 
-export const execAsync = promisify(exec);
+export const execFileAsync = promisify(execFile);
