+++ extensions/skills/src/utils/skills-cli-agents.ts
@@ -1,0 +1,86 @@
+/**
+ * Display name -> CLI agent ID for all agents supported by the Skills CLI.
+ * Sourced from https://github.com/vercel-labs/skills/blob/main/src/agents.ts
+ * Dynamically discovered agents fall back to the default transform (lowercase + space->hyphen).
+ */
+const AGENT_DISPLAY_TO_ID = new Map<string, string>([
+  ["AdaL", "adal"],
+  ["AiderDesk", "aider-desk"],
+  ["Amp", "amp"],
+  ["Antigravity", "antigravity"],
+  ["Antigravity CLI", "antigravity-cli"],
+  ["AstrBot", "astrbot"],
+  ["Autohand Code CLI", "autohand-code"],
+  ["Augment", "augment"],
+  ["Claude Code", "claude-code"],
+  ["Cline", "cline"],
+  ["Code Studio", "codestudio"],
+  ["CodeArts Agent", "codearts-agent"],
+  ["CodeBuddy", "codebuddy"],
+  ["Codemaker", "codemaker"],
+  ["Codex", "codex"],
+  ["Command Code", "command-code"],
+  ["Continue", "continue"],
+  ["Cortex Code", "cortex"],
+  ["Crush", "crush"],
+  ["Cursor", "cursor"],
+  ["Deep Agents", "deepagents"],
+  ["Devin for Terminal", "devin"],
+  ["Dexto", "dexto"],
+  ["Droid", "droid"],
+  ["Eve", "eve"],
+  ["Firebender", "firebender"],
+  ["ForgeCode", "forgecode"],
+  ["Gemini CLI", "gemini-cli"],
+  ["GitHub Copilot", "github-copilot"],
+  ["Goose", "goose"],
+  ["Hermes Agent", "hermes-agent"],
+  ["IBM Bob", "bob"],
+  ["iFlow CLI", "iflow-cli"],
+  ["inference.sh", "inference-sh"],
+  ["Jazz", "jazz"],
+  ["Junie", "junie"],
+  ["Kilo Code", "kilo"],
+  ["Kimi Code CLI", "kimi-code-cli"],
+  ["Kiro CLI", "kiro-cli"],
+  ["Kode", "kode"],
+  ["Lingma", "lingma"],
+  ["Loaf", "loaf"],
+  ["MCPJam", "mcpjam"],
+  ["Mistral Vibe", "mistral-vibe"],
+  ["Moxby", "moxby"],
+  ["Mux", "mux"],
+  ["Neovate", "neovate"],
+  ["Ona", "ona"],
+  ["OpenClaw", "openclaw"],
+  ["OpenCode", "opencode"],
+  ["OpenHands", "openhands"],
+  ["Pi", "pi"],
+  ["Pochi", "pochi"],
+  ["PromptScript", "promptscript"],
+  ["Qoder", "qoder"],
+  ["Qoder CN", "qoder-cn"],
+  ["Qwen Code", "qwen-code"],
+  ["Reasonix", "reasonix"],
+  ["Replit", "replit"],
+  ["Roo Code", "roo"],
+  ["Rovo Dev", "rovodev"],
+  ["Tabnine CLI", "tabnine-cli"],
+  ["Terramind", "terramind"],
+  ["Tinycloud", "tinycloud"],
+  ["Trae", "trae"],
+  ["Trae CN", "trae-cn"],
+  ["Universal", "universal"],
+  ["Warp", "warp"],
+  ["Windsurf", "windsurf"],
+  ["Zed", "zed"],
+  ["Zencoder", "zencoder"],
+  ["Zenflow", "zenflow"],
+]);
+
+export function agentDisplayNameToId(displayName: string): string {
+  return AGENT_DISPLAY_TO_ID.get(displayName) ?? displayName.toLowerCase().replace(/\s+/g, "-");
+}
+
+/** Sorted known agent display names, used as synchronous initial data before the CLI responds. */
+export const KNOWN_AGENT_NAMES: string[] = [...AGENT_DISPLAY_TO_ID.keys()].sort();
