+++ extensions/skills/src/utils/skills-cli.ts
@@ -1,218 +1,23 @@
-import { access, stat } from "node:fs/promises";
-import { constants } from "node:fs";
 import { homedir } from "node:os";
-import { basename } from "node:path";
-import { getCustomNpxPath, shouldDisableSkillsCliTelemetry } from "../preferences";
 import type { InstalledSkill, Skill } from "../shared";
-import { execAsync } from "./exec-async";
-import { getExecOptions } from "./exec-options";
+import { agentDisplayNameToId, KNOWN_AGENT_NAMES } from "./skills-cli-agents";
+import {
+  InvalidCustomNpxPathError,
+  NpxResolutionError,
+  isInvalidCustomNpxPathError,
+  isNpxResolutionError,
+  runSkillsCli,
+} from "./skills-cli-runner";
 
 const home = homedir();
-const isWindows = process.platform === "win32";
 
-let validatedCustomNpxPath: string | null = null;
-let pendingCustomNpxValidation: { path: string; promise: Promise<void> } | null = null;
-let pendingSkillsCliRun: Promise<unknown> = Promise.resolve();
-let bunxResolutionFailed = false;
-
-type ExecFailure = Error & {
-  code?: string | number;
-  stderr?: string;
+export {
+  InvalidCustomNpxPathError,
+  NpxResolutionError,
+  isInvalidCustomNpxPathError,
+  isNpxResolutionError,
+  KNOWN_AGENT_NAMES,
 };
-
-type PackageRunner = "npx" | "bunx";
-
-export class NpxResolutionError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "NpxResolutionError";
-  }
-}
-
-export function isNpxResolutionError(error: unknown): boolean {
-  return error instanceof NpxResolutionError;
-}
-
-export class InvalidCustomNpxPathError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "InvalidCustomNpxPathError";
-  }
-}
-
-export function isInvalidCustomNpxPathError(error: unknown): boolean {
-  return error instanceof InvalidCustomNpxPathError;
-}
-
-function buildSkillsCliCommand(runner: PackageRunner, args: string[], executable: string = runner): string {
-  const runnerArgs = runner === "npx" ? ["-y", "skills@latest"] : ["--silent", "skills@latest"];
-  return [executable, ...runnerArgs, ...args].map(shellEscape).join(" ");
-}
-
-function getSkillsCliEnvOverrides(): Record<string, string> {
-  return shouldDisableSkillsCliTelemetry() ? { DISABLE_TELEMETRY: "1" } : {};
-}
-
-async function runSkillsCli(args: string[]): Promise<string> {
-  return enqueueSkillsCliRun(() => runSkillsCliCommand(args));
-}
-
-async function enqueueSkillsCliRun<T>(run: () => Promise<T>): Promise<T> {
-  const runAfterPending = pendingSkillsCliRun.then(run, run);
-  pendingSkillsCliRun = runAfterPending.catch(() => undefined);
-  return runAfterPending;
-}
-
-async function runSkillsCliCommand(args: string[]): Promise<string> {
-  const customNpxPath = getCustomNpxPath();
-  if (customNpxPath) {
-    await validateCustomNpxPath(customNpxPath);
-
-    const execOptions = await getExecOptions();
-    try {
-      const { stdout } = await execAsync(buildSkillsCliCommand("npx", args, customNpxPath), execOptions);
-      return stdout;
-    } catch (error) {
-      throw normalizeCliError(error, customNpxPath);
-    }
-  }
-
-  if (!bunxResolutionFailed) {
-    try {
-      const execOptions = await getExecOptions();
-      const { stdout } = await execAsync(buildSkillsCliCommand("bunx", args), execOptions);
-      return stdout;
-    } catch (error) {
-      if (!isNpxCommandResolutionFailure(error, "bunx")) {
-        throw normalizeCliError(error, "bunx");
-      }
-      bunxResolutionFailed = true;
-    }
-  }
-
-  try {
-    const execOptions = await getExecOptions();
-    execOptions.env = {
-      ...execOptions.env,
-      ...getSkillsCliEnvOverrides(),
-    };
-    const { stdout } = await execAsync(buildSkillsCliCommand("npx", args), execOptions);
-    return stdout;
-  } catch (npxError) {
-    throw normalizeCliError(npxError, "npx");
-  }
-}
-
-/** Escape a value for safe use as a shell argument. */
-function shellEscape(arg: string): string {
-  if (isWindows) {
-    return `"${arg.replace(/"/g, '\\"')}"`;
-  }
-  return `'${arg.replace(/'/g, "'\\''")}'`;
-}
-
-function normalizeCliError(error: unknown, npxCommand: string): Error {
-  if (isNpxCommandResolutionFailure(error, npxCommand)) {
-    return new NpxResolutionError(
-      "Unable to find a working bunx or npx command. Install Bun, or install Node.js/npm. If you need to force a custom npx executable, set it in the extension configuration under 'Custom npx Path'.",
-    );
-  }
-
-  if (error instanceof Error) {
-    return error;
-  }
-
-  return new Error("Failed to execute the skills CLI command.");
-}
-
-async function validateCustomNpxPath(customNpxPath: string): Promise<void> {
-  if (validatedCustomNpxPath === customNpxPath) {
-    return;
-  }
-
-  if (pendingCustomNpxValidation?.path === customNpxPath) {
-    return pendingCustomNpxValidation.promise;
-  }
-
-  const validationPromise = assertValidCustomNpxPath(customNpxPath);
-  pendingCustomNpxValidation = { path: customNpxPath, promise: validationPromise };
-
-  try {
-    await validationPromise;
-    validatedCustomNpxPath = customNpxPath;
-  } finally {
-    if (pendingCustomNpxValidation?.path === customNpxPath) {
-      pendingCustomNpxValidation = null;
-    }
-  }
-}
-
-async function assertValidCustomNpxPath(customNpxPath: string): Promise<void> {
-  const invalidPathMessage =
-    "The configured Custom npx Path is incorrect. It must point to the `npx` executable. Update the path in the extension configuration or clear it to use automatic detection.";
-
-  const executableNames = isWindows ? new Set(["npx", "npx.cmd", "npx.exe"]) : new Set(["npx"]);
-  if (!executableNames.has(basename(customNpxPath).toLowerCase())) {
-    throw new InvalidCustomNpxPathError(invalidPathMessage);
-  }
-
-  let fileStats;
-  try {
-    fileStats = await stat(customNpxPath);
-  } catch {
-    throw new InvalidCustomNpxPathError(invalidPathMessage);
-  }
-  if (fileStats.isDirectory()) {
-    throw new InvalidCustomNpxPathError(invalidPathMessage);
-  }
-
-  if (!isWindows) {
-    try {
-      await access(customNpxPath, constants.X_OK);
-    } catch {
-      throw new InvalidCustomNpxPathError(invalidPathMessage);
-    }
-  }
-}
-
-function isNpxCommandResolutionFailure(error: unknown, npxCommand: string): boolean {
-  const failure = error as ExecFailure | undefined;
-  const code = typeof failure?.code === "string" || typeof failure?.code === "number" ? String(failure.code) : "";
-  const details = [failure?.message, failure?.stderr]
-    .filter((value): value is string => typeof value === "string")
-    .join("\n")
-    .toLowerCase();
-  const normalizedNpxCommand = npxCommand.toLowerCase();
-  const commandBase = basename(normalizedNpxCommand).replace(/\.(cmd|exe)$/, "");
-  const windowsCommandNotFound = `'${commandBase}' is not recognized as an internal or external command`;
-  // cmd.exe echoes the command name with its surrounding double quotes from the shell-escaped invocation,
-  // so the error reads `'"bunx"' is not recognized...` when bunx is missing.
-  const windowsCommandNotFoundQuoted = `'"${commandBase}"' is not recognized as an internal or external command`;
-
-  const mentionsCommand =
-    details.includes(`spawn ${normalizedNpxCommand} `) ||
-    details.includes(`spawn ${commandBase} `) ||
-    details.includes(`command not found: ${commandBase}`) ||
-    details.includes(`${commandBase}: command not found`) ||
-    details.includes(windowsCommandNotFound) ||
-    details.includes(windowsCommandNotFoundQuoted);
-
-  const npxShimModuleNotFound =
-    commandBase === "npx" &&
-    details.includes("cannot find module") &&
-    (details.includes("npm-prefix.js") || details.includes("npx-cli.js"));
-
-  return (
-    (code === "ENOENT" && mentionsCommand) ||
-    npxShimModuleNotFound ||
-    details.includes(`spawn ${normalizedNpxCommand} enoent`) ||
-    details.includes(`spawn ${commandBase} enoent`) ||
-    details.includes(`command not found: ${commandBase}`) ||
-    details.includes(`${commandBase}: command not found`) ||
-    details.includes(windowsCommandNotFound) ||
-    details.includes(windowsCommandNotFoundQuoted)
-  );
-}
 
 /** Shape of each entry from `skills list --json` */
 interface SkillsListJsonEntry {
@@ -253,97 +58,10 @@ export async function installSkill(skill: Skill, agentDisplayNames?: string[]): 
   await runSkillsCli(args);
 }
 
-/**
- * Display name → CLI agent ID for all agents supported by the Skills CLI.
- * Sourced from https://github.com/vercel-labs/skills/blob/main/src/agents.ts
- * Dynamically discovered agents fall back to the default transform (lowercase + space→hyphen).
- */
-const AGENT_DISPLAY_TO_ID = new Map<string, string>([
-  ["AdaL", "adal"],
-  ["AiderDesk", "aider-desk"],
-  ["Amp", "amp"],
-  ["Antigravity", "antigravity"],
-  ["Antigravity CLI", "antigravity-cli"],
-  ["AstrBot", "astrbot"],
-  ["Autohand Code CLI", "autohand-code"],
-  ["Augment", "augment"],
-  ["Claude Code", "claude-code"],
-  ["Cline", "cline"],
-  ["Code Studio", "codestudio"],
-  ["CodeArts Agent", "codearts-agent"],
-  ["CodeBuddy", "codebuddy"],
-  ["Codemaker", "codemaker"],
-  ["Codex", "codex"],
-  ["Command Code", "command-code"],
-  ["Continue", "continue"],
-  ["Cortex Code", "cortex"],
-  ["Crush", "crush"],
-  ["Cursor", "cursor"],
-  ["Deep Agents", "deepagents"],
-  ["Devin for Terminal", "devin"],
-  ["Dexto", "dexto"],
-  ["Droid", "droid"],
-  ["Eve", "eve"],
-  ["Firebender", "firebender"],
-  ["ForgeCode", "forgecode"],
-  ["Gemini CLI", "gemini-cli"],
-  ["GitHub Copilot", "github-copilot"],
-  ["Goose", "goose"],
-  ["Hermes Agent", "hermes-agent"],
-  ["IBM Bob", "bob"],
-  ["iFlow CLI", "iflow-cli"],
-  ["inference.sh", "inference-sh"],
-  ["Jazz", "jazz"],
-  ["Junie", "junie"],
-  ["Kilo Code", "kilo"],
-  ["Kimi Code CLI", "kimi-code-cli"],
-  ["Kiro CLI", "kiro-cli"],
-  ["Kode", "kode"],
-  ["Lingma", "lingma"],
-  ["Loaf", "loaf"],
-  ["MCPJam", "mcpjam"],
-  ["Mistral Vibe", "mistral-vibe"],
-  ["Moxby", "moxby"],
-  ["Mux", "mux"],
-  ["Neovate", "neovate"],
-  ["Ona", "ona"],
-  ["OpenClaw", "openclaw"],
-  ["OpenCode", "opencode"],
-  ["OpenHands", "openhands"],
-  ["Pi", "pi"],
-  ["Pochi", "pochi"],
-  ["PromptScript", "promptscript"],
-  ["Qoder", "qoder"],
-  ["Qoder CN", "qoder-cn"],
-  ["Qwen Code", "qwen-code"],
-  ["Reasonix", "reasonix"],
-  ["Replit", "replit"],
-  ["Roo Code", "roo"],
-  ["Rovo Dev", "rovodev"],
-  ["Tabnine CLI", "tabnine-cli"],
-  ["Terramind", "terramind"],
-  ["Tinycloud", "tinycloud"],
-  ["Trae", "trae"],
-  ["Trae CN", "trae-cn"],
-  ["Universal", "universal"],
-  ["Warp", "warp"],
-  ["Windsurf", "windsurf"],
-  ["Zed", "zed"],
-  ["Zencoder", "zencoder"],
-  ["Zenflow", "zenflow"],
-]);
-
-function agentDisplayNameToId(displayName: string): string {
-  return AGENT_DISPLAY_TO_ID.get(displayName) ?? displayName.toLowerCase().replace(/\s+/g, "-");
-}
-
-/** Sorted known agent display names, used as synchronous initial data before the CLI responds. */
-export const KNOWN_AGENT_NAMES: string[] = [...AGENT_DISPLAY_TO_ID.keys()].sort();
-
 export interface AgentDiscoveryResult {
   agents: string[];
   /**
-   * Maps installed skill name → agents it is installed on.
+   * Maps installed skill name -> agents it is installed on.
    * Keyed by the CLI's `name` field from `skills list --json`, which matches
    * the `skillId` used in `skills add source@skillId`.
    */
@@ -351,7 +69,7 @@ export interface AgentDiscoveryResult {
 }
 
 export async function discoverAgents(): Promise<AgentDiscoveryResult> {
-  const agentSet = new Set<string>(AGENT_DISPLAY_TO_ID.keys());
+  const agentSet = new Set<string>(KNOWN_AGENT_NAMES);
   const skillAgentMap: Record<string, string[]> = {};
   try {
     const skills = await listInstalledSkills();
@@ -376,17 +94,10 @@ export async function removeSkill(skillName: string, agentDisplayNames?: string[
   await runSkillsCli(args);
 }
 
-/**
- * Update all installed skills.
- */
 export async function updateAllSkills(): Promise<void> {
   await runSkillsCli(["update", "-g", "-y"]);
 }
 
-/**
- * Update a single installed skill by name.
- * Runs `skills update <skill-name> -g -y` via the resolved package runner (bunx with npx fallback).
- */
 export async function updateSkill(skillName: string): Promise<void> {
   await runSkillsCli(["update", skillName, "-g", "-y"]);
 }
